# openapi.yaml
openapi: 3.0.3
info:
  title: Corporate Resource Booking API
  version: 1.0.0
  description: API for booking shared corporate resources such as meeting rooms and workspaces.
servers:
  - url: http://localhost:8080
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorBody:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
    Error:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        full_name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum:
            - employee
            - admin
        created_at:
          type: string
          format: date-time
      required:
        - id
        - full_name
        - email
        - role
        - created_at
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    Resource:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        type:
          type: string
          enum:
            - meeting_room
            - workspace
        location:
          type: string
        capacity:
          type: integer
        description:
          type: string
        image_urls:
          type: array
          items:
            type: string
            format: uri
        equipment:
          type: array
          items:
            type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - type
        - location
        - capacity
        - description
        - image_urls
        - equipment
        - is_active
        - created_at
        - updated_at
    ResourceListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Resource"
      required:
        - items
...
    EquipmentListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            type: string
      required:
        - items
    Booking:
      type: object
      properties:
        id:
          type: integer
          format: int64
        resource_id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        status:
          type: string
          enum:
            - active
            - cancelled
        purpose:
          type: string
        created_at:
          type: string
          format: date-time
        cancelled_at:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - resource_id
        - user_id
        - start_time
        - end_time
        - status
        - purpose
        - created_at
    BookingListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
      required:
        - items
...
    RegisterRequest:
      type: object
      properties:
        full_name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - full_name
        - email
        - password
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    ResourceRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum:
            - meeting_room
            - workspace
        location:
          type: string
        capacity:
          type: integer
        description:
          type: string
        image_urls:
          type: array
          items:
            type: string
            format: uri
        equipment:
          type: array
          items:
            type: string
        is_active:
          type: boolean
      required:
        - name
        - type
        - location
        - capacity
        - description
    BookingRequest:
      type: object
      properties:
        resource_id:
          type: integer
          format: int64
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        purpose:
          type: string
      required:
        - resource_id
        - start_time
        - end_time
        - purpose
    ScheduleRecommendationRequest:
      type: object
      properties:
        resource_type:
          type: string
          enum:
            - meeting_room
        participants:
          type: integer
          minimum: 1
        duration_minutes:
          type: integer
          minimum: 15
          multipleOf: 15
        preferred_start:
          type: string
          format: date-time
        search_window_start:
          type: string
          format: date-time
        search_window_end:
          type: string
          format: date-time
      required:
        - resource_type
        - participants
        - duration_minutes
        - preferred_start
        - search_window_start
        - search_window_end
    ScheduleRecommendationCandidate:
      type: object
      properties:
        resource_id:
          type: integer
          format: int64
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        score:
          type: number
          format: double
        explanation:
          type: string
      required:
        - resource_id
        - start_time
        - end_time
        - score
        - explanation
    ScheduleRecommendationResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ScheduleRecommendationCandidate"
      required:
        - items
...
    UtilizationReportResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/UtilizationReportItem"
        stats:
          $ref: "#/components/schemas/UtilizationReportStats"
      required:
        - items
        - stats
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing, invalid, or expired bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated user does not have enough permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    MethodNotAllowed:
      description: Method not allowed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
paths:
...
  /auth/register:
    post:
      tags:
        - Auth
      summary: Register a new employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /auth/login:
    post:
      tags:
        - Auth
      summary: Sign in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /me:
    get:
      tags:
        - Auth
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /resources:
    get:
      tags:
        - Resources
      summary: List resources
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum:
              - meeting_room
              - workspace
        - in: query
          name: include_inactive
          schema:
            type: boolean
          description: Set to `true` to include inactive resources. Any other value is treated as `false`.
        - in: query
          name: equipment
          schema:
            type: string
          description: Repeat the parameter to require all listed equipment tags, for example `?equipment=projector&equipment=whiteboard`.
      responses:
        "200":
          description: Resource list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceListResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
    post:
      tags:
        - Resources
      summary: Create a resource
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResourceRequest"
      responses:
        "201":
          description: Resource created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /equipment:
    get:
      tags:
        - Resources
      summary: List all equipment tags
      responses:
        "200":
          description: Sorted unique equipment tags
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipmentListResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
...
  /availability:
    get:
      tags:
        - Bookings
      summary: List available resources for a time window
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: start
          required: true
          schema:
            type: string
            format: date-time
        - in: query
          name: end
          required: true
          schema:
            type: string
            format: date-time
        - in: query
          name: type
          schema:
            type: string
            enum:
              - meeting_room
              - workspace
        - in: query
          name: equipment
          schema:
            type: string
          description: Repeat the parameter to require all listed equipment tags, for example `?equipment=projector&equipment=whiteboard`.
      responses:
        "200":
          description: List of available resources
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceListResponse"
        "400":
          description: Invalid interval or availability request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /recommendations/schedule:
    post:
      tags:
        - Recommendations
      summary: Recommend the best meeting room slots inside a search window
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleRecommendationRequest"
      responses:
        "200":
          description: Ranked recommendation candidates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleRecommendationResponse"
        "400":
          description: Invalid JSON, malformed timestamps, or recommendation request failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /bookings:
    post:
      tags:
        - Bookings
      summary: Create a booking
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingRequest"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Invalid JSON, malformed timestamps, or booking creation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /bookings/my:
    get:
      tags:
        - Bookings
      summary: List current user bookings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User bookings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
...
  /admin/bookings:
    get:
      tags:
        - Admin
      summary: List all bookings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All bookings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /admin/reports/utilization:
    get:
      tags:
        - Admin
      summary: Get resource utilization report
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: start
          required: true
          schema:
            type: string
            format: date-time
        - in: query
          name: end
          required: true
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Utilization report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UtilizationReportResponse"
        "400":
          description: Invalid interval or report generation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"